requirements.txt
streamlit
pillow
requests
transformer
torch
fastapi
uvicron
python-multipart
